ReOrg utils, and move create_times from sims to utils
